Move the bolt_feature_switches table creation from UpgradeSchema to Recurring class
