+++ Setup/Recurring.php
@@ -1,0 +1,78 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+
+namespace Bolt\Boltpay\Setup;
+
+use Magento\Framework\Setup\InstallSchemaInterface;
+use Magento\Framework\Setup\ModuleContextInterface;
+use Magento\Framework\Setup\SchemaSetupInterface;
+
+/**
+ * @codeCoverageIgnore
+ */
+class Recurring implements InstallSchemaInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function install(SchemaSetupInterface $setup, ModuleContextInterface $context)
+    {
+        $setup->startSetup();
+        $this->setupFeatureSwitchTable($setup);
+        $setup->endSetup();
+    }
+
+    /**
+     * @param SchemaSetupInterface $setup
+     * @throws \Zend_Db_Exception
+     */
+    private function setupFeatureSwitchTable(SchemaSetupInterface $setup)
+    {
+        $tableCreated = $setup->getConnection()->isTableExists('bolt_feature_switches');
+
+        if ($tableCreated) {
+            return;
+        }
+
+        // Create table now that we know it doesnt already exists.
+        $table = $setup->getConnection()
+            ->newTable($setup->getTable('bolt_feature_switches'))
+            ->addColumn(
+                'id',
+                \Magento\Framework\DB\Ddl\Table::TYPE_INTEGER,
+                null,
+                ['identity' => true, 'unsigned' => true, 'nullable' => false, 'primary' => true],
+                'ID'
+            )
+            ->addColumn(
+                'switch_name',
+                \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
+                255,
+                ['nullable' => false],
+                'Switch name'
+            )
+            ->addColumn(
+                'switch_value',
+                \Magento\Framework\DB\Ddl\Table::TYPE_BOOLEAN,
+                null,
+                ['nullable' => false, 'default' => '0'],
+                'switch value'
+            )->addColumn(
+                'default_value',
+                \Magento\Framework\DB\Ddl\Table::TYPE_BOOLEAN,
+                null,
+                ['nullable' => false, 'default' => '0'],
+                'default value'
+            )->addColumn(
+                'rollout_percentage',
+                \Magento\Framework\DB\Ddl\Table::TYPE_INTEGER,
+                null,
+                ['nullable' => false, 'default' => '0'],
+                'rollout percentage'
+            )->setComment("Bolt feature switch table");
+        $setup->getConnection()->createTable($table);
+    }
+}
