+++ Setup/UpgradeSchema.php
@@ -97,8 +97,6 @@ class UpgradeSchema implements UpgradeSchemaInterface
             ]
         );
 
-        $this->setupFeatureSwitchTable($setup);
-
         $this->setUpFeatureBoltCustomerCreditCardsTable($setup);
 
         $this->setupWebhookLogTable($setup);
@@ -106,58 +104,6 @@ class UpgradeSchema implements UpgradeSchemaInterface
         $this->updateWebhookLogTable($setup);
 
         $setup->endSetup();
-    }
-
-    private function setupFeatureSwitchTable(SchemaSetupInterface $setup)
-    {
-        // If the table exists we do nothing. In the future, we can add migrations here based on
-        // current version from context, however for now just move on.
-        // The reason we do ugrade schema instead of install schema is that install schema is triggered *only*
-        // once during install. This makes debugging harder.
-        // However upgrade schema is triggered on every update and we get a chance to make changes as needed.
-        $tableCreated = $setup->getConnection()->isTableExists('bolt_feature_switches');
-
-        if ($tableCreated) {
-            return;
-        }
-
-        // Create table now that we know it doesnt already exists.
-        $table = $setup->getConnection()
-            ->newTable($setup->getTable('bolt_feature_switches'))
-            ->addColumn(
-                'id',
-                \Magento\Framework\DB\Ddl\Table::TYPE_INTEGER,
-                null,
-                ['identity' => true, 'unsigned' => true, 'nullable' => false, 'primary' => true],
-                'ID'
-            )
-            ->addColumn(
-                'switch_name',
-                \Magento\Framework\DB\Ddl\Table::TYPE_TEXT,
-                255,
-                ['nullable' => false],
-                'Switch name'
-            )
-            ->addColumn(
-                'switch_value',
-                \Magento\Framework\DB\Ddl\Table::TYPE_BOOLEAN,
-                null,
-                ['nullable' => false, 'default' => '0'],
-                'switch value'
-            )->addColumn(
-                'default_value',
-                \Magento\Framework\DB\Ddl\Table::TYPE_BOOLEAN,
-                null,
-                ['nullable' => false, 'default' => '0'],
-                'default value'
-            )->addColumn(
-                'rollout_percentage',
-                \Magento\Framework\DB\Ddl\Table::TYPE_INTEGER,
-                null,
-                ['nullable' => false, 'default' => '0'],
-                'rollout percentage'
-            )->setComment("Bolt feature switch table");
-        $setup->getConnection()->createTable($table);
     }
 
     private function setupWebhookLogTable($setup)
